add changes

+++ sql/create_tg_user.sql
@@ -6,12 +6,12 @@ CREATE TABLE `tg_user` (`id` int NOT NULL
                         , `username` varchar(255) NOT NULL
                         , `phoneNumber` varchar(255) NULL
                         , `chatId` int NOT NULL
-                        , `propertyType` smallint NOT NULL
-                        , `roomType` smallint NOT NULL
-                        , `districtType` smallint NOT NULL
-                        , `priceType` smallint NOT NULL
-                        , `apartmentPriceType` smallint NOT NULL
-                        , `menuStep` smallint NOT NULL
+                        , `propertyType` int NOT NULL
+                        , `roomType` int NOT NULL
+                        , `districtType` int NOT NULL
+                        , `priceType` int NOT NULL
+                        , `apartmentPriceType` int NOT NULL
+                        , `menuStep` int NOT NULL
                         , `isActive` tinyint NOT NULL
                         , `startTS` datetime NOT NULL
                         , `lastUpdateTS` datetime NOT NULL
